Move Execution classes to dockstore common
