+++ dockstore-common/src/main/java/io/dockstore/common/metrics/ExecutionsRequestBody.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 OICR and UCSC
+ * Copyright 2024 OICR and UCSC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,11 +15,11 @@
  *
  */
 
-package io.dockstore.webservice.core.metrics;
+package io.dockstore.common.metrics;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import io.dockstore.webservice.core.metrics.constraints.HasExecutionsOrMetrics;
-import io.dockstore.webservice.core.metrics.constraints.HasUniqueExecutionIds;
+import io.dockstore.common.metrics.constraints.HasExecutionsOrMetrics;
+import io.dockstore.common.metrics.constraints.HasUniqueExecutionIds;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
