+++ dockstore-common/src/main/java/io/dockstore/common/metrics/constraints/HasUniqueExecutionIdsValidator.java
@@ -15,9 +15,9 @@
  *
  */
 
-package io.dockstore.webservice.core.metrics.constraints;
+package io.dockstore.common.metrics.constraints;
 
-import io.dockstore.webservice.core.metrics.ExecutionsRequestBody;
+import io.dockstore.common.metrics.ExecutionsRequestBody;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import java.util.ArrayList;
