+++ dockstore-common/src/main/java/io/dockstore/common/metrics/constraints/ISO8601ExecutionDate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 OICR and UCSC
+ * Copyright 2024 OICR and UCSC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  *
  */
 
-package io.dockstore.webservice.core.metrics.constraints;
+package io.dockstore.common.metrics.constraints;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -25,16 +25,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Defines the `HasExecutions` constraint annotation, which
- * checks that ExecutionsRequestBody has executions or metrics.
+ * Defines the `ISO8601ExecutionTime` constraint annotation, which
+ * checks that the execution time is in ISO 1806 date format.
  */
 @Target({ ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE, ElementType.CONSTRUCTOR, ElementType.PARAMETER, ElementType.TYPE_USE })
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = HasExecutionsOrMetricsValidator.class)
-public @interface HasExecutionsOrMetrics {
-    String MUST_CONTAIN_EXECUTIONS_OR_METRICS = "must contain executions, like RunExecutions or ValidationExecutions, or it must contain Metrics";
+@Constraint(validatedBy = ISO8601ExecutionDateValidator.class)
+public @interface ISO8601ExecutionDate {
+    String EXECUTION_DATE_FORMAT_ERROR = "must be in ISO 8601 UTC date format";
 
-    String message() default MUST_CONTAIN_EXECUTIONS_OR_METRICS;
+    String message() default "'${validatedValue}' " + EXECUTION_DATE_FORMAT_ERROR;
     Class<?>[] groups () default {};
     Class<? extends Payload>[] payload () default {};
 }
