+++ dockstore-common/src/main/java/io/dockstore/common/metrics/constraints/ValidExecutionId.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.dockstore.webservice.core.metrics.constraints;
+package io.dockstore.common.metrics.constraints;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -25,16 +25,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Defines the `ValidClientExecutionStatus` constraint annotation, which
- * checks that the execution status is meant for client use.
+ * Defines the `ValidExecutionId` constraint annotation, which
+ * checks that execution ID is valid.
  */
 @Target({ ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE, ElementType.CONSTRUCTOR, ElementType.PARAMETER, ElementType.TYPE_USE })
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ValidClientExecutionStatusValidator.class)
-public @interface ValidClientExecutionStatus {
-    String INVALID_EXECUTION_STATUS_MESSAGE = "cannot be ALL";
+@Constraint(validatedBy = ValidExecutionIdValidator.class)
+public @interface ValidExecutionId {
+    String INVALID_EXECUTION_ID_MESSAGE = "must only consist of alphanumerics and internal underscores and hyphens, and be less than 100 characters";
 
-    String message() default INVALID_EXECUTION_STATUS_MESSAGE;
+    String message() default INVALID_EXECUTION_ID_MESSAGE;
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
 }
