+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/ExecutionStatusCountMetric.java
@@ -17,13 +17,14 @@
 
 package io.dockstore.webservice.core.metrics;
 
-import static io.dockstore.webservice.core.metrics.ExecutionStatusCountMetric.ExecutionStatus.ABORTED;
-import static io.dockstore.webservice.core.metrics.ExecutionStatusCountMetric.ExecutionStatus.FAILED;
-import static io.dockstore.webservice.core.metrics.ExecutionStatusCountMetric.ExecutionStatus.FAILED_RUNTIME_INVALID;
-import static io.dockstore.webservice.core.metrics.ExecutionStatusCountMetric.ExecutionStatus.FAILED_SEMANTIC_INVALID;
-import static io.dockstore.webservice.core.metrics.ExecutionStatusCountMetric.ExecutionStatus.SUCCESSFUL;
+import static io.dockstore.common.metrics.ExecutionStatus.ABORTED;
+import static io.dockstore.common.metrics.ExecutionStatus.FAILED;
+import static io.dockstore.common.metrics.ExecutionStatus.FAILED_RUNTIME_INVALID;
+import static io.dockstore.common.metrics.ExecutionStatus.FAILED_SEMANTIC_INVALID;
+import static io.dockstore.common.metrics.ExecutionStatus.SUCCESSFUL;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import io.dockstore.common.metrics.ExecutionStatus;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -49,7 +50,7 @@ import java.util.Map;
 @ApiModel(value = "ExecutionStatusMetric", description = "Aggregated metrics about workflow execution statuses")
 @Schema(name = "ExecutionStatusMetric", description = "Aggregated metrics about workflow execution statuses", allOf = Metric.class)
 @SuppressWarnings("checkstyle:magicnumber")
-public class ExecutionStatusCountMetric extends CountMetric<ExecutionStatusCountMetric.ExecutionStatus, MetricsByStatus> {
+public class ExecutionStatusCountMetric extends CountMetric<ExecutionStatus, MetricsByStatus> {
 
     @NotEmpty
     @MapKeyEnumerated(EnumType.STRING)
@@ -145,12 +146,4 @@ public class ExecutionStatusCountMetric extends CountMetric<ExecutionStatusCount
         return numberOfSuccessfulExecutions + numberOfFailedExecutions + numberOfAbortedExecutions;
     }
 
-    public enum ExecutionStatus {
-        ALL, // Internal use only
-        SUCCESSFUL,
-        FAILED,
-        FAILED_SEMANTIC_INVALID,
-        FAILED_RUNTIME_INVALID,
-        ABORTED
-    }
 }
