+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/ValidatorVersionInfo.java
@@ -18,7 +18,7 @@
 package io.dockstore.webservice.core.metrics;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.dockstore.webservice.core.metrics.constraints.ISO8601ExecutionDate;
+import io.dockstore.common.metrics.constraints.ISO8601ExecutionDate;
 import io.swagger.annotations.ApiModelProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.Column;
