+++ dockstore-webservice/src/main/java/io/dockstore/webservice/resources/proposedGA4GH/ToolsExtendedApiService.java
@@ -17,7 +17,7 @@ package io.dockstore.webservice.resources.proposedGA4GH;
 
 import io.dockstore.common.Partner;
 import io.dockstore.webservice.core.User;
-import io.dockstore.webservice.core.metrics.ExecutionsRequestBody;
+import io.dockstore.common.metrics.ExecutionsRequestBody;
 import io.dockstore.webservice.core.metrics.Metrics;
 import io.swagger.api.NotFoundException;
 import jakarta.ws.rs.core.MultivaluedMap;
